server/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pymysql
pydantic
python-dotenv
alembic
pydantic-settings
passlib[bcrypt]
python-jose[cryptography]
